+++ internal/discovery/discovery.go
@@ -3,7 +3,6 @@ package discovery
 import (
 	"context"
 	"fmt"
-	"hash/fnv"
 	"strings"
 
 	"google.golang.org/grpc"
@@ -71,26 +70,17 @@ func (d *client) Discover(ctx context.Context) (endpoints []endpoint.Endpoint, e
 	endpoints = make([]endpoint.Endpoint, 0, len(listEndpointsResult.Endpoints))
 	for _, e := range listEndpointsResult.Endpoints {
 		if e.Ssl == d.ssl {
-			node := endpoint.Endpoint{
-				ID:    e.NodeId,
-				Host:  e.Address,
-				Port:  int(e.Port),
-				Local: e.Location == listEndpointsResult.SelfLocation,
-			}
-			if node.ID == 0 {
-				node.ID = hash([]byte(node.Address()))
-			}
-			endpoints = append(endpoints, node)
+			endpoints = append(endpoints, endpoint.Endpoint{
+				NodeID:     e.GetNodeId(),
+				Host:       e.GetAddress(),
+				Port:       int(e.GetPort()),
+				LoadFactor: e.GetLoadFactor(),
+				Local:      e.GetLocation() == listEndpointsResult.GetSelfLocation(),
+			})
 		}
 	}
 
 	return endpoints, nil
-}
-
-func hash(b []byte) uint32 {
-	h := fnv.New32a()
-	h.Write(b)
-	return h.Sum32()
 }
 
 func (d *client) WhoAmI(ctx context.Context) (*WhoAmI, error) {